Versão final perfeita

+++ senac-cn-ado01/lib.h
@@ -9,8 +9,7 @@ typedef struct elementoDePilha{
 
 typedef enum { true, false } bool;
 
-int potencia(int n, int potencia);
-
+unsigned long long potencia(int n, int potencia);
 
 bool pilhaVazia(Celula * pilha);
 
@@ -18,12 +17,11 @@ void inserirNaPilha(Celula **pilha, char c);
 
 void printPilha(Celula *pilha,bool negativo);
 
-
 bool *  validarEntrada(char numero[],char base_entrada[], char base_saida[]);
 
-int converterParaBase10(char numero[],int base_entrada);
+unsigned long long converterParaBase10(char numero[],int base_entrada);
 
-void converterParaBaseM(unsigned long numero, int base_saida,bool negativo);
+void converterParaBaseM(unsigned long long numero, int base_saida,bool negativo);
 
 
 #endif // LIB_H_
